.gemini/requirements.txt
mcp>=1.0.0
python-dotenv>=1.0.1
asyncio>=3.4.3
google-generativeai>=0.8.3
